+++ frictionless/__init__.py
@@ -1,3 +1,5 @@
+from .actions import describe, extract, transform
+# TODO: remove in v5
 from .actions import *
 from .check import Check
 from .control import Control
@@ -27,7 +29,6 @@ from .status import Status, StatusTask
 from .step import Step
 from .storage import Storage
 from .system import system
-from .transform import *
 from .type import Type
 from .validate import *
 from . import errors
