+++ frictionless/actions/__init__.py
@@ -1,6 +1,8 @@
 from .describe import describe
 from .extract import extract
+from .transform import transform
 
 # TODO: remove these legacy imports in v5
 from .describe import describe_dialect, describe_resource, describe_package, describe_schema
 from .extract import extract_resource, extract_package
+from .transform import transform_resource, transform_package, transform_pipeline
