+++ frictionless/plugins/server/server.py
@@ -1,7 +1,6 @@
 import multiprocessing
-from ...actions import describe, extract
+from ...actions import describe, extract, transform
 from ...validate import validate
-from ...transform import transform
 from ...server import Server
 from ... import helpers
 from ... import settings
