--- frictionless/transform/__init__.py
@@ -1,4 +1,0 @@
-from .main import transform
-from .package import transform_package
-from .pipeline import transform_pipeline
-from .resource import transform_resource
